pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "pydantic==2.9.2",
    "pyswisseph==2.10.3.2",
    "python-dotenv==1.0.1",
    "httpx==0.27.2",
    "openai==2.13.0",
    "tzdata==2024.2",
    "asyncpg==0.29.0",
    "requests==2.32.3",
]
